Make plugin update commands available in Agents window
