+++ src/vs/workbench/contrib/chat/browser/agentPluginsView.ts
@@ -10,7 +10,7 @@ import { IListContextMenuEvent } from '../../../../base/browser/ui/list/list.js'
 import { IPagedRenderer } from '../../../../base/browser/ui/list/listPaging.js';
 import { Action, IAction, Separator } from '../../../../base/common/actions.js';
 import { RunOnceScheduler } from '../../../../base/common/async.js';
-import { CancellationToken, CancellationTokenSource } from '../../../../base/common/cancellation.js';
+import { CancellationTokenSource } from '../../../../base/common/cancellation.js';
 import { Codicon } from '../../../../base/common/codicons.js';
 import { getErrorMessage } from '../../../../base/common/errors.js';
 import { Event } from '../../../../base/common/event.js';
@@ -22,7 +22,7 @@ import { dirname } from '../../../../base/common/resources.js';
 import { localize, localize2 } from '../../../../nls.js';
 import { Action2, MenuId, registerAction2 } from '../../../../platform/actions/common/actions.js';
 import { IConfigurationService } from '../../../../platform/configuration/common/configuration.js';
-import { ContextKeyExpr, IContextKey, IContextKeyService } from '../../../../platform/contextkey/common/contextkey.js';
+import { ContextKeyExpr, IContextKeyService } from '../../../../platform/contextkey/common/contextkey.js';
 import { IContextMenuService } from '../../../../platform/contextview/browser/contextView.js';
 import { IHoverService } from '../../../../platform/hover/browser/hover.js';
 import { SyncDescriptor } from '../../../../platform/instantiation/common/descriptors.js';
@@ -53,7 +53,7 @@ import { hasSourceChanged, IMarketplacePlugin, IPluginMarketplaceService } from 
 import { AgentPluginEditorInput } from './agentPluginEditor/agentPluginEditorInput.js';
 import { AgentPluginItemKind, IAgentPluginItem, IInstalledPluginItem, IMarketplacePluginItem } from './agentPluginEditor/agentPluginItems.js';
 import { getInstalledPluginContextMenuActions, InstallPluginAction, OpenPluginReadmeAction } from './agentPluginActions.js';
-import { ForceUpdateAgentPluginsCommandId, HasInstalledAgentPluginsContext, InstalledAgentPluginsViewId, RefreshAgentPluginMarketplacesCommandId, UpdateAgentPluginsCommandId, UpdatingAgentPluginsContext } from './chat.js';
+import { HasInstalledAgentPluginsContext, InstalledAgentPluginsViewId, RefreshAgentPluginMarketplacesCommandId } from './chat.js';
 
 //#region Item model
 
@@ -559,39 +559,6 @@ export class AgentPluginsListView extends AbstractExtensionsListView<IAgentPlugi
 
 //#region Browse command
 
-let updatingPluginsContextKey: IContextKey<boolean> | undefined;
-let updatePluginsPromise: Promise<void> | undefined;
-
-function updatePlugins(accessor: ServicesAccessor, force: boolean): Promise<void> {
-	if (updatePluginsPromise) {
-		return updatePluginsPromise;
-	}
-
-	// Services must be resolved synchronously — the accessor is invalidated as
-	// soon as the command handler returns, so a lookup after the `await` below
-	// would throw instead of showing the notification.
-	const pluginInstallService = accessor.get(IPluginInstallService);
-	const notificationService = accessor.get(INotificationService);
-
-	updatingPluginsContextKey?.set(true);
-	updatePluginsPromise = (async () => {
-		try {
-			const result = await pluginInstallService.updateAllPlugins({ force }, CancellationToken.None);
-			if (result.updatedNames.length === 0 && result.failedNames.length === 0) {
-				notificationService.info(localize('agentPlugins.upToDate', "Plugins are up to date."));
-			}
-		} catch (error) {
-			notificationService.error(localize('agentPlugins.updateFailed', "Failed to update plugins: {0}", getErrorMessage(error)));
-			throw error;
-		} finally {
-			updatePluginsPromise = undefined;
-			updatingPluginsContextKey?.set(false);
-		}
-	})();
-
-	return updatePluginsPromise;
-}
-
 class AgentPluginsBrowseCommand extends Action2 {
 	constructor() {
 		super({
@@ -615,55 +582,6 @@ class AgentPluginsBrowseCommand extends Action2 {
 
 	async run(accessor: ServicesAccessor) {
 		accessor.get(IExtensionsWorkbenchService).openSearch('@agentPlugins ');
-	}
-}
-
-class CheckForPluginUpdatesCommand extends Action2 {
-	constructor() {
-		super({
-			id: UpdateAgentPluginsCommandId,
-			title: localize2('agentPlugins.checkForUpdates', "Update Plugins"),
-			category: localize2('chat.category', "Chat"),
-			icon: Codicon.refresh,
-			precondition: ContextKeyExpr.and(ChatContextKeys.enabled, UpdatingAgentPluginsContext.negate()),
-			f1: true,
-			menu: [{
-				id: MenuId.ViewTitle,
-				when: ContextKeyExpr.and(
-					ContextKeyExpr.equals('view', InstalledAgentPluginsViewId),
-					ChatContextKeys.Setup.hidden.negate(),
-					ChatContextKeys.Setup.disabledInWorkspace.negate(),
-				),
-				group: 'navigation',
-				order: 1,
-				alt: {
-					id: ForceUpdateAgentPluginsCommandId,
-					title: localize2('agentPlugins.forceUpdate', "Update Plugins (Force)"),
-					icon: Codicon.refresh,
-				},
-			}],
-		});
-	}
-
-	async run(accessor: ServicesAccessor) {
-		await updatePlugins(accessor, false);
-	}
-}
-
-class ForceUpdatePluginsCommand extends Action2 {
-	constructor() {
-		super({
-			id: ForceUpdateAgentPluginsCommandId,
-			title: localize2('agentPlugins.forceUpdate', "Update Plugins (Force)"),
-			category: localize2('chat.category', "Chat"),
-			icon: Codicon.refresh,
-			precondition: ContextKeyExpr.and(ChatContextKeys.enabled, UpdatingAgentPluginsContext.negate()),
-			f1: true,
-		});
-	}
-
-	async run(accessor: ServicesAccessor) {
-		await updatePlugins(accessor, true);
 	}
 }
 
@@ -736,14 +654,11 @@ export class AgentPluginsViewsContribution extends Disposable implements IWorkbe
 		super();
 
 		const hasInstalledKey = HasInstalledAgentPluginsContext.bindTo(contextKeyService);
-		updatingPluginsContextKey = UpdatingAgentPluginsContext.bindTo(contextKeyService);
 		this._register(autorun(reader => {
 			hasInstalledKey.set(agentPluginService.plugins.read(reader).length > 0);
 		}));
 
 		registerAction2(AgentPluginsBrowseCommand);
-		registerAction2(CheckForPluginUpdatesCommand);
-		registerAction2(ForceUpdatePluginsCommand);
 		registerAction2(RefreshPluginMarketplacesCommand);
 
 		Registry.as<IViewsRegistry>(ViewExtensions.ViewsRegistry).registerViews([
